fix: rts version file moved to src
